+++ libPolyLine/Line.c
@@ -3,14 +3,14 @@
 #include <V2.h>
 #include <math.h>
 
-long double Line_Evaluate(Line l, V2 v2)
+double Line_Evaluate(Line l, V2 v2)
 {
     return l.a * v2.x + l.b * v2.y + l.c;
 }
 
 Line Line_Normalize(Line l)
 {
-    long double m = sqrtl(l.a * l.a + l.b * l.b);
+    double m = sqrtl(l.a * l.a + l.b * l.b);
 
     Line result =
     {
@@ -24,7 +24,7 @@ Line Line_Normalize(Line l)
 
 V2 Line_IntersectionLinePointDirection(Line l, V2 point, V2 direction)
 {
-    long double t =
+    double t =
         (-1.0 / (l.a * direction.x + l.b * direction.y)) * (l.c + l.a * point.x + l.b * point.y);
 
     V2 result =
@@ -39,7 +39,7 @@ V2 Line_IntersectionLinePointDirection(Line l, V2 point, V2 direction)
 
 V2 Line_PointProjectionOnNormalizedLine(Line l, V2 point)
 {
-    long double t = -1.0 * (l.c + l.a * point.x + l.b * point.y);
+    double t = -1.0 * (l.c + l.a * point.x + l.b * point.y);
 
     V2 result =
     {
